first spilt of stats functions
